requirements.txt
flask==2.3.3
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0